Since we can have module-specific conftest files, moved the fixtures that are used only by the config variables tests into a new one in data_context
